[model-gateway] refactor: unify worker management into modular workflow structure
Restructure worker lifecycle management into a unified module:

- Extract shared steps (RegisterWorkersStep, UpdatePoliciesStep, ActivateWorkersStep)
- Organize local worker steps (registration + removal) under worker/local/
- Organize external worker steps under worker/external/
- Move workflow builders to respective module files
- Maintain backward compatibility with deprecated type aliases

Module structure:
  worker/
  ├── shared/     - Steps shared across workflows
  ├── local/      - SGLang/vLLM worker management (registration, removal)
  └── external/   - External API worker registration
