+++ sgl-model-gateway/src/core/steps/mod.rs
@@ -1,26 +1,62 @@
 //! Workflow step implementations
 //!
 //! This module contains concrete step implementations for various workflows:
-//! - Local worker registration (SGLang, vLLM - gRPC or HTTP modes)
-//! - External worker registration (OpenAI, xAI, Anthropic, etc. - HTTPS only)
-//! - Worker removal
+//! - Worker management (registration, removal, future: updates)
 //! - MCP server registration
 //! - WASM module registration and removal
 //! - Future: Tokenizer fetching, LoRA updates, etc.
 
-pub mod external_worker_registration;
 pub mod mcp_registration;
 pub mod wasm_module_registration;
 pub mod wasm_module_removal;
-pub mod worker_registration;
-pub mod worker_removal;
+pub mod worker;
 
-pub use external_worker_registration::{
-    create_external_worker_registration_workflow, group_models_into_cards,
-    infer_model_type_from_id, ActivateExternalWorkersStep, CreateExternalWorkersStep,
-    DiscoverModelsStep, ModelInfo, ModelsResponse, RegisterExternalWorkersStep,
-    UpdateExternalPoliciesStep,
+// Worker management (registration, removal)
+#[allow(deprecated)]
+pub use worker::create_external_worker_registration_workflow;
+// Backward compatibility aliases
+#[allow(deprecated)]
+pub use worker::create_worker_registration_workflow;
+pub use worker::{
+    // Workflow builders
+    create_external_worker_workflow,
+    create_local_worker_workflow,
+    create_worker_removal_workflow,
+    // Utility functions
+    group_models_into_cards,
+    infer_model_type_from_id,
+    // Shared steps
+    ActivateWorkersStep,
+    // External registration steps
+    CreateExternalWorkersStep,
+    // Local registration steps
+    CreateLocalWorkerStep,
+    DetectConnectionModeStep,
+    DiscoverDPInfoStep,
+    DiscoverMetadataStep,
+    DiscoverModelsStep,
+    DpInfo,
+    // Removal steps
+    FindWorkersToRemoveStep,
+    ModelInfo,
+    ModelsResponse,
+    RegisterWorkersStep,
+    RemoveFromPolicyRegistryStep,
+    RemoveFromWorkerRegistryStep,
+    UpdatePoliciesStep,
+    UpdateRemainingPoliciesStep,
+    WorkerList,
+    WorkerRemovalRequest,
 };
+
+// Legacy type aliases for backward compatibility
+pub type ActivateWorkerStep = ActivateWorkersStep;
+pub type RegisterWorkerStep = RegisterWorkersStep;
+pub type CreateWorkerStep = CreateLocalWorkerStep;
+pub type ActivateExternalWorkersStep = ActivateWorkersStep;
+pub type RegisterExternalWorkersStep = RegisterWorkersStep;
+pub type UpdateExternalPoliciesStep = UpdatePoliciesStep;
+
 pub use mcp_registration::{
     create_mcp_registration_workflow, ConnectMcpServerStep, DiscoverMcpInventoryStep,
     McpServerConfigRequest, RegisterMcpServerStep, ValidateRegistrationStep,
@@ -33,13 +69,4 @@ pub use wasm_module_registration::{
 pub use wasm_module_removal::{
     create_wasm_module_removal_workflow, FindModuleToRemoveStep, RemoveModuleStep,
     WasmModuleRemovalRequest,
-};
-pub use worker_registration::{
-    create_worker_registration_workflow, ActivateWorkerStep, CreateWorkerStep,
-    DetectConnectionModeStep, DiscoverDPInfoStep, DiscoverMetadataStep, RegisterWorkerStep,
-    UpdatePoliciesStep,
-};
-pub use worker_removal::{
-    create_worker_removal_workflow, FindWorkersToRemoveStep, RemoveFromPolicyRegistryStep,
-    RemoveFromWorkerRegistryStep, UpdateRemainingPoliciesStep, WorkerRemovalRequest,
 };
