+++ sgl-model-gateway/src/core/steps/worker/local/detect_connection.rs
@@ -1,0 +1,146 @@
+//! Connection mode detection step.
+
+use std::{sync::Arc, time::Duration};
+
+use async_trait::async_trait;
+use reqwest::Client;
+use tracing::debug;
+
+use crate::{
+    app_context::AppContext,
+    core::ConnectionMode,
+    protocols::worker_spec::WorkerConfigRequest,
+    routers::grpc::client::GrpcClient,
+    workflow::{StepExecutor, StepId, StepResult, WorkflowContext, WorkflowError, WorkflowResult},
+};
+
+/// Strip protocol prefix from URL.
+fn strip_protocol(url: &str) -> String {
+    url.trim_start_matches("http://")
+        .trim_start_matches("https://")
+        .trim_start_matches("grpc://")
+        .to_string()
+}
+
+/// Try HTTP health check.
+async fn try_http_health_check(
+    url: &str,
+    timeout_secs: u64,
+    client: &Client,
+) -> Result<(), String> {
+    let is_https = url.starts_with("https://");
+    let protocol = if is_https { "https" } else { "http" };
+    let clean_url = strip_protocol(url);
+    let health_url = format!("{}://{}/health", protocol, clean_url);
+
+    client
+        .get(&health_url)
+        .timeout(Duration::from_secs(timeout_secs))
+        .send()
+        .await
+        .and_then(reqwest::Response::error_for_status)
+        .map_err(|e| format!("Health check failed: {}", e))?;
+
+    Ok(())
+}
+
+/// Perform gRPC health check with runtime type.
+async fn do_grpc_health_check(
+    grpc_url: &str,
+    timeout_secs: u64,
+    runtime_type: &str,
+) -> Result<(), String> {
+    let connect_future = GrpcClient::connect(grpc_url, runtime_type);
+    let client = tokio::time::timeout(Duration::from_secs(timeout_secs), connect_future)
+        .await
+        .map_err(|_| "gRPC connection timeout".to_string())?
+        .map_err(|e| format!("gRPC connection failed: {}", e))?;
+
+    let health_future = client.health_check();
+    tokio::time::timeout(Duration::from_secs(timeout_secs), health_future)
+        .await
+        .map_err(|_| "gRPC health check timeout".to_string())?
+        .map_err(|e| format!("gRPC health check failed: {}", e))?;
+
+    Ok(())
+}
+
+/// Try gRPC health check (tries SGLang first, then vLLM if not specified).
+async fn try_grpc_health_check(
+    url: &str,
+    timeout_secs: u64,
+    runtime_type: Option<&str>,
+) -> Result<(), String> {
+    let grpc_url = if url.starts_with("grpc://") {
+        url.to_string()
+    } else {
+        format!("grpc://{}", strip_protocol(url))
+    };
+
+    match runtime_type {
+        Some(runtime) => do_grpc_health_check(&grpc_url, timeout_secs, runtime).await,
+        None => {
+            // Try SGLang first, then vLLM as fallback
+            if let Ok(()) = do_grpc_health_check(&grpc_url, timeout_secs, "sglang").await {
+                return Ok(());
+            }
+            do_grpc_health_check(&grpc_url, timeout_secs, "vllm")
+                .await
+                .map_err(|e| format!("gRPC failed (tried SGLang and vLLM): {}", e))
+        }
+    }
+}
+
+/// Step 1: Detect connection mode by probing HTTP and gRPC.
+pub struct DetectConnectionModeStep;
+
+#[async_trait]
+impl StepExecutor for DetectConnectionModeStep {
+    async fn execute(&self, context: &mut WorkflowContext) -> WorkflowResult<StepResult> {
+        let config: Arc<WorkerConfigRequest> = context.get_or_err("worker_config")?;
+        let app_context: Arc<AppContext> = context.get_or_err("app_context")?;
+
+        debug!(
+            "Detecting connection mode for {} (timeout: {}s, max_attempts: {})",
+            config.url, config.health_check_timeout_secs, config.max_connection_attempts
+        );
+
+        // Try both protocols in parallel
+        let url = config.url.clone();
+        let timeout = config.health_check_timeout_secs;
+        let client = &app_context.client;
+        let runtime_type = config.runtime.as_deref();
+
+        let (http_result, grpc_result) = tokio::join!(
+            try_http_health_check(&url, timeout, client),
+            try_grpc_health_check(&url, timeout, runtime_type)
+        );
+
+        let connection_mode = match (http_result, grpc_result) {
+            (Ok(_), _) => {
+                debug!("{} detected as HTTP", config.url);
+                ConnectionMode::Http
+            }
+            (_, Ok(_)) => {
+                debug!("{} detected as gRPC", config.url);
+                ConnectionMode::Grpc { port: None }
+            }
+            (Err(http_err), Err(grpc_err)) => {
+                return Err(WorkflowError::StepFailed {
+                    step_id: StepId::new("detect_connection_mode"),
+                    message: format!(
+                        "Both HTTP and gRPC health checks failed for {}: HTTP: {}, gRPC: {}",
+                        config.url, http_err, grpc_err
+                    ),
+                });
+            }
+        };
+
+        context.set("connection_mode", connection_mode);
+        Ok(StepResult::Success)
+    }
+
+    fn is_retryable(&self, _error: &WorkflowError) -> bool {
+        true
+    }
+}
