+++ sgl-model-gateway/src/core/steps/worker/local/discover_dp.rs
@@ -1,0 +1,79 @@
+//! Data Parallel (DP) information discovery step.
+
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use tracing::debug;
+
+use super::discover_metadata::get_server_info;
+use crate::{
+    protocols::worker_spec::WorkerConfigRequest,
+    workflow::{StepExecutor, StepId, StepResult, WorkflowContext, WorkflowError, WorkflowResult},
+};
+
+/// DP (Data Parallel) information for a worker.
+#[derive(Debug, Clone)]
+pub struct DpInfo {
+    pub dp_size: usize,
+    pub model_id: String,
+}
+
+/// Get DP info for a worker URL.
+pub async fn get_dp_info(url: &str, api_key: Option<&str>) -> Result<DpInfo, String> {
+    let info = get_server_info(url, api_key).await?;
+
+    let dp_size = info
+        .dp_size
+        .ok_or_else(|| format!("No dp_size in response from {}", url))?;
+
+    let model_id = info
+        .model_id
+        .filter(|s| !s.is_empty())
+        .or(info.served_model_name.filter(|s| !s.is_empty()))
+        .or_else(|| {
+            info.model_path
+                .and_then(|path| path.split('/').next_back().map(|s| s.to_string()))
+        })
+        .unwrap_or_else(|| "unknown".to_string());
+
+    Ok(DpInfo { dp_size, model_id })
+}
+
+/// Step 2b: Discover DP (Data Parallel) information (only for DP-aware workers).
+pub struct DiscoverDPInfoStep;
+
+#[async_trait]
+impl StepExecutor for DiscoverDPInfoStep {
+    async fn execute(&self, context: &mut WorkflowContext) -> WorkflowResult<StepResult> {
+        let config: Arc<WorkerConfigRequest> = context.get_or_err("worker_config")?;
+
+        if !config.dp_aware {
+            debug!(
+                "Worker {} is not DP-aware, skipping DP discovery",
+                config.url
+            );
+            return Ok(StepResult::Success);
+        }
+
+        debug!("Discovering DP info for {} (DP-aware)", config.url);
+
+        let dp_info = get_dp_info(&config.url, config.api_key.as_deref())
+            .await
+            .map_err(|e| WorkflowError::StepFailed {
+                step_id: StepId::new("discover_dp_info"),
+                message: format!("Failed to get DP info: {}", e),
+            })?;
+
+        debug!(
+            "Discovered DP size {} for {} (model: {})",
+            dp_info.dp_size, config.url, dp_info.model_id
+        );
+
+        context.set("dp_info", dp_info);
+        Ok(StepResult::Success)
+    }
+
+    fn is_retryable(&self, _error: &WorkflowError) -> bool {
+        true
+    }
+}
