+++ sgl-model-gateway/src/core/steps/worker/local/mod.rs
@@ -1,0 +1,224 @@
+//! Local worker management steps for SGLang/vLLM workers.
+//!
+//! This module handles the lifecycle of workers that run locally
+//! or on accessible infrastructure with SGLang/vLLM runtimes:
+//! - **Registration**: Adding new local workers
+//! - **Removal**: Removing local workers
+
+mod create_worker;
+mod detect_connection;
+mod discover_dp;
+mod discover_metadata;
+mod removal;
+
+use std::{sync::Arc, time::Duration};
+
+pub use create_worker::CreateLocalWorkerStep;
+pub use detect_connection::DetectConnectionModeStep;
+pub use discover_dp::{get_dp_info, DiscoverDPInfoStep, DpInfo};
+pub use discover_metadata::DiscoverMetadataStep;
+pub use removal::{
+    FindWorkersToRemoveStep, RemoveFromPolicyRegistryStep, RemoveFromWorkerRegistryStep,
+    UpdateRemainingPoliciesStep, WorkerRemovalRequest,
+};
+
+use super::shared::{ActivateWorkersStep, RegisterWorkersStep, UpdatePoliciesStep};
+use crate::{
+    config::RouterConfig,
+    workflow::{BackoffStrategy, FailureAction, RetryPolicy, StepDefinition, WorkflowDefinition},
+};
+
+/// Create local worker registration workflow definition.
+///
+/// DAG structure with parallel execution opportunities:
+/// ```text
+///                detect_connection_mode
+///                         │
+///            ┌────────────┴────────────┐
+///            │                         │
+///    discover_metadata         discover_dp_info
+///            │                         │
+///            └────────────┬────────────┘
+///                         │
+///                   create_worker
+///                         │
+///                  register_workers
+///                         │
+///            ┌────────────┴────────────┐
+///            │                         │
+///     update_policies          activate_workers
+///            │                         │
+///            └────────────┴────────────┘
+/// ```
+pub fn create_local_worker_workflow(router_config: &RouterConfig) -> WorkflowDefinition {
+    let detect_timeout = Duration::from_secs(router_config.worker_startup_timeout_secs);
+
+    // Calculate max_attempts based on timeout
+    let timeout_secs = detect_timeout.as_secs() as f64;
+    let effective_timeout = timeout_secs * 0.9;
+    let max_attempts = if effective_timeout > 10.0 {
+        (5 + ((effective_timeout - 10.0) / 5.0).ceil() as u32).max(3)
+    } else {
+        3
+    };
+
+    WorkflowDefinition::new("local_worker_registration", "Local Worker Registration")
+        // Step 1: Detect connection mode (HTTP vs gRPC)
+        .add_step(
+            StepDefinition::new(
+                "detect_connection_mode",
+                "Detect Connection Mode",
+                Arc::new(DetectConnectionModeStep),
+            )
+            .with_retry(RetryPolicy {
+                max_attempts,
+                backoff: BackoffStrategy::Linear {
+                    increment: Duration::from_secs(1),
+                    max: Duration::from_secs(5),
+                },
+            })
+            .with_timeout(detect_timeout)
+            .with_failure_action(FailureAction::FailWorkflow),
+        )
+        // Step 2a: Discover metadata (parallel with DP discovery)
+        .add_step(
+            StepDefinition::new(
+                "discover_metadata",
+                "Discover Metadata",
+                Arc::new(DiscoverMetadataStep),
+            )
+            .with_retry(RetryPolicy {
+                max_attempts: 3,
+                backoff: BackoffStrategy::Fixed(Duration::from_secs(1)),
+            })
+            .with_timeout(Duration::from_secs(10))
+            .with_failure_action(FailureAction::ContinueNextStep)
+            .depends_on(&["detect_connection_mode"]),
+        )
+        // Step 2b: Discover DP info (parallel with metadata discovery)
+        .add_step(
+            StepDefinition::new(
+                "discover_dp_info",
+                "Discover DP Info",
+                Arc::new(DiscoverDPInfoStep),
+            )
+            .with_retry(RetryPolicy {
+                max_attempts: 3,
+                backoff: BackoffStrategy::Fixed(Duration::from_secs(1)),
+            })
+            .with_timeout(Duration::from_secs(10))
+            .with_failure_action(FailureAction::FailWorkflow)
+            .depends_on(&["detect_connection_mode"]),
+        )
+        // Step 3: Create worker(s)
+        .add_step(
+            StepDefinition::new(
+                "create_worker",
+                "Create Worker",
+                Arc::new(CreateLocalWorkerStep),
+            )
+            .with_timeout(Duration::from_secs(5))
+            .with_failure_action(FailureAction::FailWorkflow)
+            .depends_on(&["discover_metadata", "discover_dp_info"]),
+        )
+        // Step 4: Register workers (shared step)
+        .add_step(
+            StepDefinition::new(
+                "register_workers",
+                "Register Workers",
+                Arc::new(RegisterWorkersStep),
+            )
+            .with_timeout(Duration::from_secs(5))
+            .with_failure_action(FailureAction::FailWorkflow)
+            .depends_on(&["create_worker"]),
+        )
+        // Step 5a: Update policies (parallel with activation)
+        .add_step(
+            StepDefinition::new(
+                "update_policies",
+                "Update Policies",
+                Arc::new(UpdatePoliciesStep),
+            )
+            .with_timeout(Duration::from_secs(5))
+            .with_failure_action(FailureAction::ContinueNextStep)
+            .depends_on(&["register_workers"]),
+        )
+        // Step 5b: Activate workers (parallel with policy update)
+        .add_step(
+            StepDefinition::new(
+                "activate_workers",
+                "Activate Workers",
+                Arc::new(ActivateWorkersStep),
+            )
+            .with_timeout(Duration::from_secs(5))
+            .with_failure_action(FailureAction::FailWorkflow)
+            .depends_on(&["register_workers"]),
+        )
+}
+
+/// Create a worker removal workflow definition.
+///
+/// DAG structure:
+/// ```text
+///     find_workers_to_remove
+///              │
+///     remove_from_policy_registry
+///              │
+///     remove_from_worker_registry
+///              │
+///     update_remaining_policies
+/// ```
+pub fn create_worker_removal_workflow() -> WorkflowDefinition {
+    WorkflowDefinition::new("worker_removal", "Remove worker from router")
+        .add_step(
+            StepDefinition::new(
+                "find_workers_to_remove",
+                "Find workers to remove",
+                Arc::new(FindWorkersToRemoveStep),
+            )
+            .with_timeout(Duration::from_secs(10))
+            .with_retry(RetryPolicy {
+                max_attempts: 1,
+                backoff: BackoffStrategy::Fixed(Duration::from_secs(0)),
+            }),
+        )
+        .add_step(
+            StepDefinition::new(
+                "remove_from_policy_registry",
+                "Remove workers from policy registry",
+                Arc::new(RemoveFromPolicyRegistryStep),
+            )
+            .with_timeout(Duration::from_secs(10))
+            .with_retry(RetryPolicy {
+                max_attempts: 1,
+                backoff: BackoffStrategy::Fixed(Duration::from_secs(0)),
+            })
+            .depends_on(&["find_workers_to_remove"]),
+        )
+        .add_step(
+            StepDefinition::new(
+                "remove_from_worker_registry",
+                "Remove workers from worker registry",
+                Arc::new(RemoveFromWorkerRegistryStep),
+            )
+            .with_timeout(Duration::from_secs(10))
+            .with_retry(RetryPolicy {
+                max_attempts: 1,
+                backoff: BackoffStrategy::Fixed(Duration::from_secs(0)),
+            })
+            .depends_on(&["remove_from_policy_registry"]),
+        )
+        .add_step(
+            StepDefinition::new(
+                "update_remaining_policies",
+                "Update cache-aware policies for remaining workers",
+                Arc::new(UpdateRemainingPoliciesStep),
+            )
+            .with_timeout(Duration::from_secs(10))
+            .with_retry(RetryPolicy {
+                max_attempts: 1,
+                backoff: BackoffStrategy::Fixed(Duration::from_secs(0)),
+            })
+            .depends_on(&["remove_from_worker_registry"]),
+        )
+}
